third_party/WebKit/LayoutTests/animations/compounding-transitions-on-repeatable-list.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#target {
  transition-property: stroke-dasharray;
  transition-duration: 1s;
  transition-delay: -0.5s;
}
</style>
<div id="target"></div>
<script>
test(() => {
  const iterations = 13;
  for (let i = 0; i < iterations; ++i) {
    target.style.strokeDasharray = '1 '.repeat(i);
    const animatedValueLength = getComputedStyle(target).strokeDasharray.split(' ').length;
    assert_less_than_equal(animatedValueLength, 1000, 'iteration ' + i);
  }
}, 'Starting compounded transitions on a repeatable list property should limit how much the animated value gets repeated to runaway memory usage.');
</script>
